Static channels from the config (public and private)

+++ plugin/src/main/java/com/craftmend/openaudiomc/spigot/modules/voicechat/channels/Channel.java
@@ -7,6 +7,7 @@ import com.craftmend.openaudiomc.generic.storage.enums.StorageKey;
 import com.craftmend.openaudiomc.generic.user.User;
 import com.craftmend.openaudiomc.spigot.modules.voicechat.VoiceChannelService;
 import lombok.Getter;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.Collection;
 import java.util.HashMap;
@@ -15,15 +16,61 @@ import java.util.UUID;
 
 public class Channel {
 
+    @Nullable
     @Getter private User creator;
     @Getter private String name;
-    private Map<UUID, ClientConnection> members = new HashMap<>();
-    private VoiceChannelService voiceChannelService;
+    private final Map<UUID, ClientConnection> members = new HashMap<>();
+    private final VoiceChannelService voiceChannelService;
+    private final String requiredPermission;
+    @Getter private final ChannelType type;
 
-    public Channel(User creator, String name, VoiceChannelService voiceChannelService) {
+    public Channel(@Nullable User creator, String name, VoiceChannelService voiceChannelService) {
         this.creator = creator;
         this.name = name;
         this.voiceChannelService = voiceChannelService;
+        this.requiredPermission = null;
+        this.type = ChannelType.USER_CHANNEL;
+    }
+
+    public Channel(String name, String permission, VoiceChannelService voiceChannelService) {
+        this.name = name;
+        this.voiceChannelService = voiceChannelService;
+        this.requiredPermission = permission;
+        type = permission != null ? ChannelType.STATIC_CHANNEL_LOCKED : ChannelType.STATIC_CHANNEL;
+    }
+
+    public ChannelEnterResponse attemptEnter(User member) {
+        // admin check
+        if (member.hasPermission("openaudiomc.voice.channel.join") || member.isAdministrator()) {
+            return ChannelEnterResponse.OK;
+        }
+
+        switch (type) {
+            case USER_CHANNEL:
+                return ChannelEnterResponse.INVITE_ONLY;
+            case STATIC_CHANNEL:
+            case STATIC_CHANNEL_LOCKED:
+                if (requiredPermission == null || member.hasPermission(requiredPermission)) {
+                    return ChannelEnterResponse.OK;
+                } else {
+                    return ChannelEnterResponse.NO_PERMISSION;
+                }
+        }
+
+        // should never happen
+        return ChannelEnterResponse.NO_PERMISSION;
+    }
+
+    public String getReadableType() {
+        switch (type) {
+            case USER_CHANNEL:
+                return StorageKey.MESSAGE_VOICE_CHANNEL_TYPE_USER.getString();
+            case STATIC_CHANNEL:
+                return StorageKey.MESSAGE_VOICE_CHANNEL_TYPE_STATIC.getString();
+            case STATIC_CHANNEL_LOCKED:
+                return StorageKey.MESSAGE_VOICE_CHANNEL_TYPE_STATIC_LOCKED.getString();
+        }
+        return "Unknown";
     }
 
     public void addMember(User user) {
@@ -62,6 +109,7 @@ public class Channel {
     }
 
     private void checkAbandonment() {
+        if (type != ChannelType.USER_CHANNEL) return;
         if (members.size() < 2) {
             members.forEach((uuid, clientConnection) -> clientConnection.getUser().sendMessage(Platform.translateColors(StorageKey.MESSAGE_VOICE_CHANNEL_ABANDONED.getString())));
             voiceChannelService.deleteChannel(name);
